Cleanup

+++ src/main/java/nl/mvdr/adventofcode/adventofcode2017/day08/Condition.java
@@ -30,8 +30,8 @@ class Condition implements Predicate<Map<String, Integer>> {
     
     @Override
     public boolean test(Map<String, Integer> registers) {
-        int registerValue = registers.getOrDefault(registerName, Integer.valueOf(0)).intValue();
-        return operator.test(Integer.valueOf(registerValue), Integer.valueOf(value));
+        Integer registerValue = registers.getOrDefault(registerName, Integer.valueOf(0));
+        return operator.test(registerValue, Integer.valueOf(value));
     }
     
     @Override
